+++ src/rendering/react/components/DnsZone/RecordComponent.ts
@@ -1,0 +1,3 @@
+import { FC } from "react";
+
+export type RecordComponent<T> = FC<{ record: T }>;
