+++ src/rendering/react/components/DnsZone/RecordSetManagedByMittwald.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import { Text } from "ink";
+
+export const RecordSetManagedByMittwald = () => (
+  <Text color="green">managed by mittwald</Text>
+);
