+++ src/rendering/react/components/DnsZone/RecordSetValues.tsx
@@ -1,0 +1,6 @@
+import React, { FC } from "react";
+import { Text } from "ink";
+
+export const RecordSetValues: FC<{ records: string[] }> = ({ records }) => {
+  return <Text>{records.join("\n")}</Text>;
+};
